requirements.txt
fastapi
uvicorn
PyYAML
psycopg2-binary
pytest
httpx

# Auth dependencies
PyJWT
passlib[bcrypt]
